backend/requirements.txt
amqp==5.0.9
APScheduler==3.8.1
asgiref==3.5.0
billiard==3.6.4.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.11.0
django-crontab==0.7.1
djangorestframework==3.13.1
kombu==5.2.3
prompt-toolkit==3.0.28
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
vine==5.0.0
wcwidth==0.2.5
